splitted up logic/FN in multiple files for better structuring and reusability
